+++ .npmignore
@@ -1,5 +1,7 @@
-src
-examples
+src/
+examples/
 .babelrc
 webpack.config.js
-public
+node_modules/
+package-lock.json
+tsconfig.json
