+++ package.json
@@ -1,34 +1,19 @@
 {
   "name": "react-world-countries-map",
-  "version": "1.0.1",
+  "version": "1.0.5",
   "private": false,
   "displayName": "React World Map of Countries",
-  "main": "./lib/index.tsx",
-  "description": "A pure react component to visualize world maps with highlighted countries. Simple. Free. No Registration, No access token, No back-end service.",
+  "main": "./dist/index.",
+  "description": "A pure react component to visualize world maps with highlighted countries. Simple. Free.",
   "authors": [
     "Yanivam (https://github.com/yanivam)"
   ],
   "dependencies": {
-    "@types/d3-geo": "^1.11.1",
-    "@types/immutable": "^3.8.7",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.25",
-    "@types/react-dom": "^16.9.5",
-    "@types/react-map-gl": "^5.2.0",
-    "@types/webpack": "^4.41.8",
-    "babel-core": "^6.21.0",
-    "d3-geo": "^1.11.9",
-    "immutable": "^4.0.0-rc.12",
-    "path": "^0.12.7",
     "prop-types": "^15.6.0",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1",
-    "react-svg-tooltip": "0.0.11",
-    "typescript": "~3.7.2",
-    "webpack": "^4.42.1"
+    "react-dom": "^16.13.1"
   },
+  "type": "webrtc",
   "peerDependencies": {
     "prop-types": "^15.6.0",
     "react": "^16.0.0",
@@ -53,6 +38,22 @@
     ]
   },
   "devDependencies": {
+    "@types/d3-geo": "^1.11.1",
+    "@types/immutable": "^3.8.7",
+    "@types/jest": "^24.0.0",
+    "@types/node": "^12.0.0",
+    "@types/react": "^16.9.25",
+    "@types/react-dom": "^16.9.5",
+    "@types/react-map-gl": "^5.2.0",
+    "@types/webpack": "^4.41.8",
+    "babel-core": "^6.21.0",
+    "d3-geo": "^1.11.9",
+    "immutable": "^4.0.0-rc.12",
+    "path": "^0.12.7",
+    "react-scripts": "3.4.1",
+    "react-svg-tooltip": "0.0.11",
+    "typescript": "~3.7.2",
+    "webpack": "^4.42.1",
     "@babel/core": "^7.9.0",
     "babel-loader": "^8.1.0",
     "gulp-babel": "^8.0.0",
