--- public/manifest.json
@@ -1,8 +1,0 @@
-{
-  "short_name": "Sample App",
-  "name": "App Sample",
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
