+++ tsconfig.json
@@ -1,6 +1,8 @@
 {
   "compilerOptions": {
+    "outDir": "./dist",
     "target": "es5",
+    "declaration": true,
     "lib": [
       "dom",
       "dom.iterable",
@@ -13,6 +15,8 @@
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "module": "esnext",
+    "noImplicitAny": true,
+    "removeComments": true,
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
@@ -20,5 +24,8 @@
   },
   "include": [
     "src"
+  ],
+  "exclude": [
+    "node_modules"
   ]
 }
